Move UnmannedSystem to a separate module to be able to add Mavlink Unmanned System under "Sensors" as an individual sensor driver.
