+++ other-modules/comm/sensorhub-comm-mavlink2/src/main/java/org/sensorhub/impl/comm/mavlink2/UnmannedSwarm.java
@@ -10,7 +10,7 @@ public class UnmannedSwarm extends SensorSystem {
         getMembers();
 
         // Add control input that distributes LLA command among members
-        addControlInput();
+        //addControlInput();
     }
 
 }
